backend/pyproject.toml
[project]
name = "my-diary"
version = "0.1.0"
description = "Spiritual practice + wellness journaling backend."
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "PyJWT[crypto]>=2.8",
    "bcrypt>=4.1",
    "python-multipart>=0.0.12",
    "structlog>=24.4",
    "email-validator>=2.2",
    "aiosqlite>=0.20",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "black>=24.10",
    "ruff>=0.7",
    "mypy>=1.13",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_diary"]
